% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulamed.R
\name{lista_remedios}
\alias{lista_remedios}
\title{Lista remédios por letra}
\usage{
lista_remedios(l)
}
\arguments{
\item{l}{letra}
}
\value{
data.frame com nome do remédio e url p/ bula
}
\description{
Lista remédios por letra
}

