% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulamed.R
\name{lista_todos_remedios}
\alias{lista_todos_remedios}
\title{Pegar nome de todos os remédios}
\usage{
lista_todos_remedios(progress = T)
}
\value{
lista de todos os remédios
}
\description{
Pegar nome de todos os remédios
}
\note{
essa função procura remédios com todas as letras do alfabeto.
}

