% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulamed.R
\name{texto_bula}
\alias{texto_bula}
\alias{texto_bulas}
\title{Texto Bula}
\usage{
texto_bula(url)

texto_bulas(url, time = 0.5, progress = T)
}
\arguments{
\item{url}{url p/ a bula do remédio}

\item{time}{tempo entre cada requisição p/ não bloquear}

\item{progress}{indica se você quer ter uma barra de progresso}
}
\value{
texto inteiro da bula do remédio
}
\description{
Texto Bula

Texto de vários urls
}
\note{
o texto pode conter partes indesejadas, como avisos do site
em que foi feito webscrapping.
}

