% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulamed.R
\name{url_letra}
\alias{url_letra}
\title{URL da lista de remédios por letra}
\usage{
url_letra(l)
}
\arguments{
\item{l}{letra}
}
\value{
url da lista de remédios dada a letra
}
\description{
URL da lista de remédios por letra
}

